Align dimension column dtypes with SQL schema and remove dead code
- Narrow integer keys from int64 to int32 across all dimensions:
  GeographyKey, HouseholdKey, CustomerKey, StoreKey, PromotionKey,
  CurrencyKey, PlanKey, SubscriptionKey, ReturnReasonKey,
  FromCurrencyKey, ToCurrencyKey
- Narrow BIT-schema flags from int32/int64 to int8: IsCurrent, IsActive,
  SalariedFlag, SalesPersonFlag, OpenFlag, IsPrimary, HasFreeTrial,
  IsActiveFlag
- Narrow VersionNumber from int64 to int8 and update static_schemas.py
  to TINYINT for both Customers and Products tables
- Narrow SMALLINT-schema columns to int16: SalesChannelKey, SortOrder,
  TypicalFulfillmentDays, AssignmentSequence
- Narrow TierRank from int32 to int8 to match TINYINT schema
- Fix Mexico continent from "South America" to "North America" in
  geography fallback data
- Replace TypeError with ConfigError in geography normalizer to match
  the project exception hierarchy
- Fix ISO column default from True to False in dates/columns.py to
  match config_schema.py
- Fix "Non-Working day" capitalization to "Non-Working Day"
- Remove dead expand_subscription_periods function (67 lines) replaced
  by vectorized generate_subscriptions_bulk, plus 9 associated tests
- Remove dead geography reload and iterrows loop in customer parallel
  orchestrator (results were never used downstream)

+++ src/dimensions/employees/transfers.py
@@ -314,7 +314,7 @@ def apply_transfers(
                 "EndDate": new_end,
                 "FTE": col_fte[idx],
                 "RoleAtStore": col_role[idx],
-                "IsPrimary": np.int32(1),
+                "IsPrimary": np.int8(1),
                 "TransferReason": str(reason),
                 "Status": new_status,
             })
